+++ data-raw/packages.R
@@ -35,7 +35,8 @@ paletteer_packages <- tibble::tribble(
   "wesanderson", "karthik/wesanderson", "0.3.6.9000", TRUE, "0.3.6",
   "yarrr", "ndphillips/yarrr", "0.1.6", TRUE, "0.1.5",
   "lisa", "tyluRp/lisa", "0.1.1.9000", TRUE, "0.1.1",
-  "IslamicArt", "lambdamoses/IslamicArt", "0.1.0", FALSE, NA
+  "IslamicArt", "lambdamoses/IslamicArt", "0.1.0", FALSE, NA,
+  "basetheme", "KKPMW/basetheme", "0.1.2", TRUE, "0.1.2"
 ) %>%
   arrange(Name)